chore: moved json logger to utilities crate - removed logger.rs file
